added piece selection class, added multiple marker differentiation
Added class for selection, which will use its own model and marker for
the user interaction.
This also required modification of the marker detection. There are now
two markers being searched for (and string values have also been passed
as names), then when the results are detected, they are all iterated
through and compared to the name strings to determine which marker is in
view.