nuance not working as anticipated, switch to speech.recognition
nuance's voice recogntion is really low, like 20% or less which is not
viable for the app and I cannot build a grammer without paying a
significant amount of money so moot.  There is now a project called ARVR
that has a custom grammar embedded.  The app now works with it, you just
need to set a ip address for the computer running the server.